Jwt Authentication done

+++ Backend/HealthCareSystem/src/main/java/com/healthcare/dto/Appointment.java
@@ -18,7 +18,7 @@ public class Appointment {
 
     @Id
     @GeneratedValue
-    private BigInteger appointmentId;
+    private Integer appointmentId;
 
     @OneToOne (cascade = CascadeType.ALL)
     private User user;
@@ -34,4 +34,3 @@ public class Appointment {
     private boolean approved;
 
 }
-
